# apps/web/ui/lib/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EnableRequest": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "EnableRequest",
        "type": "object"
      },
...
      "ExtensionSourceRequest": {
        "properties": {
          "catalog_root": {
            "default": "",
            "title": "Catalog Root",
            "type": "string"
          },
          "extension_id": {
            "default": "",
            "title": "Extension Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "ref": {
            "default": "",
            "title": "Ref",
            "type": "string"
          },
          "sha256": {
            "default": "",
            "title": "Sha256",
            "type": "string"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          },
          "version": {
            "default": "",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ExtensionSourceRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InstallRequest": {
        "properties": {
          "confirmations": {
            "items": {
              "type": "string"
            },
            "title": "Confirmations",
            "type": "array"
          },
          "preview_id": {
            "default": "",
            "title": "Preview Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ExtensionSourceRequest"
          }
        },
        "required": [
          "source"
        ],
        "title": "InstallRequest",
        "type": "object"
      },
...
      "InvokeRequest": {
        "properties": {
          "command_type": {
            "title": "Command Type",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "command_type"
        ],
        "title": "InvokeRequest",
        "type": "object"
      },
...
      "PreviewRequest": {
        "properties": {
          "extension_id": {
            "default": "",
            "title": "Extension Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ExtensionSourceRequest"
          }
        },
        "required": [
          "source"
        ],
        "title": "PreviewRequest",
        "type": "object"
      },
...
      "RollbackRequest": {
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "RollbackRequest",
        "type": "object"
      },
...
      "UninstallRequest": {
        "properties": {
          "confirm": {
            "default": false,
            "title": "Confirm",
            "type": "boolean"
          },
          "preserve_artifacts": {
            "default": true,
            "title": "Preserve Artifacts",
            "type": "boolean"
          },
          "preserve_logs": {
            "default": true,
            "title": "Preserve Logs",
            "type": "boolean"
          },
          "preserve_state": {
            "default": true,
            "title": "Preserve State",
            "type": "boolean"
          }
        },
        "title": "UninstallRequest",
        "type": "object"
      },
      "UpgradeRequest": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "confirmations": {
            "items": {
              "type": "string"
            },
            "title": "Confirmations",
            "type": "array"
          },
          "preview_id": {
            "default": "",
            "title": "Preview Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ExtensionSourceRequest"
          }
        },
        "required": [
          "source"
        ],
        "title": "UpgradeRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Project Muteki — Agent 安全工作台",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/extensions/install": {
      "post": {
        "operationId": "install_api_extensions_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Install Api Extensions Install Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install"
      }
    },
...
    "/api/extensions/preview": {
      "post": {
        "operationId": "preview_api_extensions_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Preview Api Extensions Preview Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview"
      }
    },
    "/api/extensions/{extension_id}": {
      "delete": {
        "operationId": "uninstall_api_extensions__extension_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/UninstallRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Uninstall Api Extensions  Extension Id  Delete"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uninstall"
      },
      "get": {
        "operationId": "get_extension_api_extensions__extension_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Extension Api Extensions  Extension Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Extension"
      }
    },
...
    "/api/extensions/{extension_id}/disable": {
      "post": {
        "operationId": "disable_api_extensions__extension_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Disable Api Extensions  Extension Id  Disable Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable"
      }
    },
    "/api/extensions/{extension_id}/enable": {
      "post": {
        "operationId": "enable_api_extensions__extension_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Enable Api Extensions  Extension Id  Enable Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable"
      }
    },
...
    "/api/extensions/{extension_id}/invoke": {
      "post": {
        "operationId": "invoke_api_extensions__extension_id__invoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Invoke Api Extensions  Extension Id  Invoke Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invoke"
      }
    },
    "/api/extensions/{extension_id}/logs": {
      "get": {
        "operationId": "extension_logs_api_extensions__extension_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Extension Logs Api Extensions  Extension Id  Logs Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extension Logs"
      }
    },
...
    "/api/extensions/{extension_id}/projection": {
      "get": {
        "operationId": "extension_projection_api_extensions__extension_id__projection_get",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "default": "",
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Extension Projection Api Extensions  Extension Id  Projection Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extension Projection"
      }
    },
    "/api/extensions/{extension_id}/rollback": {
      "post": {
        "operationId": "rollback_api_extensions__extension_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Rollback Api Extensions  Extension Id  Rollback Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rollback"
      }
    },
...
    "/api/extensions/{extension_id}/upgrade": {
      "post": {
        "operationId": "upgrade_api_extensions__extension_id__upgrade_post",
        "parameters": [
          {
            "in": "path",
            "name": "extension_id",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Upgrade Api Extensions  Extension Id  Upgrade Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upgrade"
      }
    },
...
  }
}